requirements.txt
altgraph==0.16.1
aniso8601==8.0.0
astroid==2.3.2
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
Click==7.0
configparser==4.0.2
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
Flask-RESTful==0.3.7
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
public==2019.4.13
pycodestyle==2.5.0
Pygments==2.4.2
PyInstaller==3.5
pylint==2.4.3
pypi-xmlrpc==2019.4.13
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
